resources/requirements.txt
packaging==25.0
python-can==4.6.1
typing_extensions==4.15.0
wrapt==1.17.3
